doc: Corrected denoising data type conversion

+++ src/algorithms/denoising/_avg_filter.py
@@ -67,4 +67,4 @@ def avg_filter(
     src = cv2.cvtColor(src.cpu().numpy().astype(np.uint8), cv2.COLOR_RGB2BGR)
 
     out = cv2.blur(src, ksize, borderType=borderType)
-    return torch.from_numpy(cv2.cvtColor(out, cv2.COLOR_BGR2RGB).astype(float) / 255.0)
+    return torch.from_numpy(cv2.cvtColor(out, cv2.COLOR_BGR2RGB).astype(np.float32) / 255.0)
